allow float values for weights

+++ frontend/configurable-ranking/src/Body.js
@@ -50,7 +50,7 @@ const Body = () => {
             let val = await cur.current.value;
             // console.log(val);
             if (isNaN(val)) return;
-            newWeights[i] = parseInt(val);
+            newWeights[i] = parseFloat(val);
         }
         setWeights(newWeights);
     };
